Move dao/repo/mappers and related entities to channel package
