+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/database/ChatDatabase.kt
@@ -6,8 +6,6 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import androidx.sqlite.db.SupportSQLiteDatabase
-import io.getstream.chat.android.livedata.dao.ChannelDao
-import io.getstream.chat.android.livedata.entity.ChannelEntity
 import io.getstream.chat.android.livedata.repository.database.converter.DateConverter
 import io.getstream.chat.android.livedata.repository.database.converter.ExtraDataConverter
 import io.getstream.chat.android.livedata.repository.database.converter.FilterObjectConverter
@@ -15,6 +13,8 @@ import io.getstream.chat.android.livedata.repository.database.converter.ListConv
 import io.getstream.chat.android.livedata.repository.database.converter.MapConverter
 import io.getstream.chat.android.livedata.repository.database.converter.SetConverter
 import io.getstream.chat.android.livedata.repository.database.converter.SyncStatusConverter
+import io.getstream.chat.android.livedata.repository.domain.channel.ChannelDao
+import io.getstream.chat.android.livedata.repository.domain.channel.ChannelEntity
 import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigDao
 import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigInnerEntity
 import io.getstream.chat.android.livedata.repository.domain.channelconfig.CommandInnerEntity
