+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/domain/channel/ChannelRepository.kt
@@ -1,12 +1,9 @@
-package io.getstream.chat.android.livedata.repository
+package io.getstream.chat.android.livedata.repository.domain.channel
 
 import androidx.collection.LruCache
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.livedata.dao.ChannelDao
-import io.getstream.chat.android.livedata.repository.mapper.toEntity
-import io.getstream.chat.android.livedata.repository.mapper.toModel
 import java.util.Date
 
 internal interface ChannelRepository {
