+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/domain/channel/userread/ChannelUserReadEntity.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.livedata.entity
+package io.getstream.chat.android.livedata.repository.domain.channel.userread
 
 import java.util.Date
