+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/repository/helper/BaseRepositoryFacadeTest.kt
@@ -2,8 +2,8 @@ package io.getstream.chat.android.livedata.repository.helper
 
 import androidx.annotation.CallSuper
 import com.nhaarman.mockitokotlin2.mock
-import io.getstream.chat.android.livedata.repository.ChannelRepository
 import io.getstream.chat.android.livedata.repository.RepositoryFacade
+import io.getstream.chat.android.livedata.repository.domain.channel.ChannelRepository
 import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigRepository
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepository
 import io.getstream.chat.android.livedata.repository.domain.queryChannels.QueryChannelsRepository
